requirements.txt
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
Flask-MySQLdb==0.2.0
get==2019.4.13
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jsonify==0.5
jwt==0.6.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
post==2019.4.13
public==2019.4.13
pycparser==2.19
PyJWT==1.6.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
PyYAML==5.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
